modified from int to int32_t

+++ main/Src/crc.c
@@ -17,7 +17,7 @@
 /* VARIABLES */
 
 /* FUNCTION PROTOTYPE */
-static uint32_t reflect(uint32_t data, int bits);
+static uint32_t reflect(uint32_t data, uint8_t bits);
 
 /**
  * @brief Reflects (reverses) the lower 'bits' number of bits in the given data.
@@ -30,9 +30,9 @@ static uint32_t reflect(uint32_t data, int bits);
  *
  * @return      The reflected value of the input data over the specified number of bits.
  */
-static uint32_t reflect(uint32_t data, int bits) {
+static uint32_t reflect(uint32_t data, uint8_t bits) {
     uint32_t reflection = 0;
-    for (int i = 0; i < bits; i++) {
+    for (uint8_t i = 0; i < bits; i++) {
         if (data & (1 << i)) {
             reflection |= (1 << (bits - 1 - i));
         }
@@ -60,7 +60,7 @@ uint32_t crc32(const uint8_t *data, uint32_t length) {
         uint8_t byte = reflect(data[i], 8);
         crc ^= ((uint32_t)byte) << 24;
 
-        for (int j = 0; j < 8; j++) {
+        for (uint8_t j = 0; j < 8; j++) {
             if (crc & 0x80000000)
                 crc = (crc << 1) ^ poly;
             else
